refactor(edge/stacks): move edge stacks components to module
