--- app/edge/EdgeDevices/EdgeDevicesView/index.ts
@@ -1,1 +1,0 @@
-export { EdgeDevicesView } from './EdgeDevicesView';
