+++ app/edge/react/components/index.ts
@@ -2,10 +2,33 @@ import angular from 'angular';
 
 import { EdgeGroupsSelector } from '@/react/edge/components/EdgeGroupsSelector';
 import { r2a } from '@/react-tools/react2angular';
+import { withReactQuery } from '@/react-tools/withReactQuery';
+import { EdgeCheckinIntervalField } from '@/react/edge/components/EdgeCheckInIntervalField';
+import { EdgeScriptForm } from '@/react/edge/components/EdgeScriptForm';
 
 export const componentsModule = angular
   .module('portainer.edge.react.components', [])
   .component(
     'edgeGroupsSelector',
     r2a(EdgeGroupsSelector, ['items', 'onChange', 'value'])
+  )
+  .component(
+    'edgeScriptForm',
+    r2a(withReactQuery(EdgeScriptForm), [
+      'edgeInfo',
+      'commands',
+      'isNomadTokenVisible',
+    ])
+  )
+  .component(
+    'edgeCheckinIntervalField',
+    r2a(withReactQuery(EdgeCheckinIntervalField), [
+      'value',
+      'onChange',
+      'isDefaultHidden',
+      'tooltip',
+      'label',
+      'readonly',
+      'size',
+    ])
   ).name;
