+++ app/edge/react/index.ts
@@ -1,9 +1,17 @@
 import angular from 'angular';
 
+import { r2a } from '@/react-tools/react2angular';
+import { withCurrentUser } from '@/react-tools/withCurrentUser';
+import { withReactQuery } from '@/react-tools/withReactQuery';
+import { withUIRouter } from '@/react-tools/withUIRouter';
+import { WaitingRoomView } from '@/react/edge/edge-devices/WaitingRoomView';
+
 import { componentsModule } from './components';
 import { viewsModule } from './views';
 
-export const reactModule = angular.module('portainer.edge.react', [
-  viewsModule,
-  componentsModule,
-]).name;
+export const reactModule = angular
+  .module('portainer.edge.react', [viewsModule, componentsModule])
+  .component(
+    'waitingRoomView',
+    r2a(withUIRouter(withReactQuery(withCurrentUser(WaitingRoomView))), [])
+  ).name;
