+++ app/react/edge/components/useIntervalOptions.ts
