+++ app/react/edge/edge-devices/ListView/EdgeDevicesDatatable/AMTDevicesDatatable/index.ts
@@ -1,0 +1,1 @@
+export { AMTDevicesDatatable } from './AMTDevicesDatatable';
