+++ app/react/edge/edge-devices/ListView/EdgeDevicesDatatable/columns/RowContext.tsx
