+++ app/react/edge/edge-devices/ListView/EdgeDevicesDatatable/columns/actions.tsx
