+++ app/react/edge/edge-devices/ListView/EdgeDevicesDatatable/columns/heartbeat.tsx
