+++ app/react/edge/edge-devices/ListView/EdgeDevicesDatatable/columns/index.tsx
