+++ app/react/edge/edge-devices/ListView/EdgeDevicesDatatable/columns/name.tsx
