+++ app/react/edge/edge-devices/ListView/index.ts
@@ -1,0 +1,1 @@
+export { ListView } from './ListView';
