+++ app/react/edge/edge-devices/WaitingRoomView/Datatable/types.ts
