+++ app/react/edge/edge-devices/WaitingRoomView/WaitingRoomView.tsx
