+++ app/react/edge/edge-devices/WaitingRoomView/index.ts
