+++ app/react/edge/edge-devices/WaitingRoomView/queries.ts
